Changed Soma Potential Slider Function
Soma Potential Slider now represents reset potential and firing
threshold, i.e. the soma fires, if the orange end of the slider is
reached (max_threshold) and reset to the reset potential/blue end of
the slider. The difference of the potential above the threshold is then
added to the minimum value.